Change label on new admin view to username
The new admin screen that comes up initially upon booting the app labels
the first field 'Name', but the subsequent login form asks for 'Username'.
This was a point of confusion for me. Changed the label on the new admin
view to reference the same label translation as the login form.