Updated tests to acommodate differences in MySQL
As detailed in:
https://github.com/mischa78/boxroom/issues/42

MySQL does not reset the auto-increment counter when a transaction is
rolled back, resulting in some of the tests failing when changing over
to MySQL because the @root_folder instance variable in Folder holds on
to the original root folder created in the first test. Added
Folder#clear_root! method to clear/reset the root folder in between
tests. Also changed ShareLink.emails field to validate max length of 255
instead of 256, as MySQL strings are limited to 255 characters.